set statements support for cursor
